Fix handling of multiple possible episode groups
In some cases, it is possible that there is more than one group of episodes
totalling the same length as the main feature (case in point was Silent Witness,
which has options to play all episodes, both parts of an episode, or each part of each
episode. This creates 2 groups that match the length of the play all options.
This was theorietically supported, but had never been tested before, so had bugs.